Bug: Doesn't generate thumbnails or crashes in some cases - fixed
Crash:
Description: Crash happens when application was started first time in the device(.json
file doesn't excist). In gridView it reads metadata and generates the thumbnails but doesn't
save them in file. When user plays a video program checks for .json file
and doesn't find it so it starts to generate thumbnails(in the thread)
again. If in that moment user presses back button, application will
crash(it is connected with QtCore library). However if user presses back
button after generation everything will be OK.
-fixed by saving thumbnail URL after its's generation

Doesn't generate thumbnails:
-If application was closed during thumbnail generation, next time it
starts, it will not generating thumbnails for the videos with blank
thumbnail.
-fixed by checking which videos don't have thumbanils