MACADAMIAN-NUTTYPLAYER-342
This is a temporary fix. For farther the metadata reader and file completness check sould be tuned to detect file copy or camera video capture situations.
The observer will wait for 5 secs before sending the file for metadata read to higher chances that it will be ready at the moment.